Changed the structure of the posts entities to use named keys for the friendly url title, rather than a specific field for it. That way, those url titles are unique.

+++ war/WEB-INF/groovy/admin/postDelete.groovy
@@ -1,6 +1,6 @@
 import com.google.appengine.api.datastore.Key
 
-def key = ['posts', params.id.toLong()] as Key
+def key = ['posts', params.id] as Key
 key.delete()
 
 redirect '/admin/posts'
